Change segLength and numPoints to prepare for using images for the chain
then fix the bugs that change uncovered 😅 

Turns out having the gravity set so high and the constraintIterations so low, made the chain stretch out real far and I didn't realize it because I couldn't see the whole thing. These changes put the full chain on screen for now to help get the sim working. Later I'll need to move it back up and set its length and boundaries appropriately for the game

This commit also takes care of some garbage collection issues that were causing the game to completely freeze for half a second every 2 or 3 seconds